Fixed bug in BuddyAllocator in Reserver method. Code adopted to new Linux-based build environment.
git-svn-id: svn://127.0.0.1@17 6511da6e-0655-744e-9e86-4527dd187af8

+++ mainline/kernel/mem/mm.cpp
@@ -3,7 +3,7 @@
  * $Id$
  *
  * This file is a part of PhobOS operating system.
- * Copyright ©AST 2009. Written by Artemy Lebedev.
+ * Copyright ï¿½AST 2009. Written by Artemy Lebedev.
  */
 
 #include <sys.h>
@@ -174,13 +174,13 @@ MM::VtoP(vaddr_t va)
 }
 
 void *
-operator new(u32 size)
+operator new(size_t size)
 {
 	return MM::OpNew(size);
 }
 
 void *
-operator new[](u32 size)
+operator new[](size_t size)
 {
 	return MM::OpNew(size);
 }
